requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.4
django-filter==22.1
flower==0.9.7
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.36
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
vine==1.3.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.2
